Fixed in_network function - returned false results
https://github.com/tehmaze/ipcalc/issues/19
https://github.com/tehmaze/ipcalc/issues/4

The return is now better to read. The rule for detecting if an IP
address is within the network is:
Address <= Testing IP <= Broadcast

This function has only the ability to check an IP address by assuming
the netmask as /32 (IPv4) or /128 (Ipv6). If the tested IP has a
netmask greater than our network it will probably fail.