New function - check_collision(), in_network() deprecated
The new function check_collisions() needs an IP address as argument. It
can be a single address or a complete network. It checks the networks
against each other and returns true if any collisions will be found.
This function is an improvement of the function in_network(). So,
in_network is now marked as deprecated and prints a warning when being
used.